create coin gecko source

+++ backend/api/DTOs/CurrencyDtos/AddCurrencyDto.cs
@@ -3,7 +3,7 @@ namespace api.DTOs.CurrencyDtos;
 public record AddCurrencyDto(
     string Symbol,
     string FullName,
-    float CurrencyPrice,
+    Decimal CurrencyPrice,
     float MarketCap,
     string Category,
     string Status,
